expense-tracker-be/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.1
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
billiard==4.2.4
Booktype==1.5
CacheControl==0.14.4
cachetools==6.2.3
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
Django==5.2.11
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.121.3
filetype==1.2.0
firebase_admin==7.1.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-generativeai==0.8.4
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-genai==2.0.8
langchain-text-splitters==0.3.5
langgraph==0.2.60
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74
langsmith==0.2.11
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
pandas==2.2.3
passlib==1.7.4
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.3.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
simplejson==3.20.2
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.5
starlette==0.50.0
tenacity==9.1.2
tomli==2.3.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.6.0
yarl==1.22.0
